+++ pscraper/api/__init__.py
@@ -2,6 +2,10 @@ from pscraper.utils.base_api import BaseAPI
 
 
 class API(BaseAPI):
+    """
+    Provides the APIs to interact with the database
+    """
+
     def __init__(self, username, password, localhost=False):
         url = 'http://pscraper.herokuapp.com/pscraper/' if not localhost else 'http://localhost:8000/pscraper/'
         super().__init__(url, (username, password))